10/30 10:00

+++ solver/solver.hpp
@@ -37,7 +37,7 @@ int node_id1;         //接続されているFPGAid
 int node_id2;         //接続されているFPGAid
 
 long cost;   //TDM比
-vector<pair<long, long> > used_net; //枝を使ったネットid first:id second :cost
+vector<pair<int, long> > used_net; //枝を使ったネットid first:id second :cost
 double sum; //各ネットの1/TDMの合計
 
 edge() : cost(0) {
@@ -64,7 +64,7 @@ vector<int> target_sig;         //送信先node
 
 long priority; //ルーティングの優先順位、高い方から先にルーティングする
 long cost;   //最終的なネットのTDM
-vector<pair<long, long> > T;   //ダイクストラ用の最短経路edgeid first:edgeid,second:TDM
+vector<pair<int, long> > T;   //ダイクストラ用の最短経路edgeid first:edgeid,second:TDM
 
 bool max; //スコアとなるグループに属している場合true
 int max_his; //過去何回かで何回以上最大グループに属している場合正の値
